Add files via upload
About the project :

The United States has more mass shootings than any other country in the world. The
Congressional Research Service defines a mass shooting as a “ shooting in which a gunman:

*  Kills four or more people
* selects victims randomly
* attacks in public place ”

Recent events, like the Las Vegas mass shooting which claimed 59 lives with over 500 injured, and the Texas Church shooting that claimed 26 lives with ~cc 20 injured, make it necessary to analyze mass shootings and derive insights from it for prediction and prevention. 

Keeping in mind the importance of mass shootings as an issue in the United States, our project aims to tackle this problem from a data analysis point of view.

What the project involves : 

Predictive and Forecast Analysis of USA Mass Shooting Dataset - Data Preprocessing/Cleaning, Data exploration, classification and  time series analysis.

The project was implemented in Jupyter notebook in Python. 

Datasets Used : 

1. https://www.kaggle.com/carlosparadis/stanford-msa/data
2. https://www.kaggle.com/gsnehaa21/federal-holidays-usa-19662020/data